renamed sample/ to samples/; refactored simple_exchange sample
