+++ samples/basic-js/environment/src/app/bandit.js
