+++ samples/basic-js/environment/src/app/training.js
