+++ samples/basic-js/environment/src/entry.js
