+++ samples/basic-js/environment/src/index.js
