+++ samples/basic-js/environment/src/lib/client.js
