+++ samples/basic-js/environment/src/lib/logging.js
