+++ samples/basic-js/environment/src/lib/wspipe.js
