+++ samples/basic-js/webpack.config.js
