+++ samples/simple-exchange-js/environment/src/app/training.js
