+++ samples/simple-exchange-js/environment/src/entry.js
