+++ samples/simple-exchange-js/environment/src/lib/client.js
