+++ samples/simple-exchange-js/environment/src/lib/logging.js
