+++ samples/simple-exchange-js/environment/src/lib/wspipe.js
