+++ samples/simple-exchange-js/webpack.config.js
